requirements.txt
openai
streamlit
manim
numpy
pillow
requests
python-dotenv
